create SwiftMockConversation
