+++ Wire-iOS Tests/ConversationDetail/GroupDetailsViewControllerSnapshotTests.swift
@@ -19,40 +19,6 @@
 import XCTest
 @testable import Wire
 
-final class MockGroupDetailsConversation: NSObject, GroupDetailsConversationType {
-    var freeParticipantSlots: Int = 1
-
-    var isUnderLegalHold: Bool = false
-    var isSelfAnActiveMember: Bool = true
-    var userDefinedName: String?
-
-    var displayName: String = ""
-
-    var sortedOtherParticipants: [UserType] = []
-    var sortedServiceUsers: [UserType] = []
-
-    var securityLevel: ZMConversationSecurityLevel = .notSecure
-
-    var allowGuests: Bool = false
-    var hasReadReceiptsEnabled: Bool = false
-
-    var conversationType: ZMConversationType = .group
-    var mutedMessageTypes: MutedMessageTypes = .none
-
-    var teamRemoteIdentifier: UUID?
-    var team: Team?
-
-    func localParticipantsContain(user: UserType) -> Bool {
-        return false
-    }
-    
-    var accessMode: ConversationAccessMode?
-    var accessRole: ConversationAccessRole?
-
-    var connectedUserType: UserType?
-    var messageDestructionTimeout: WireDataModel.MessageDestructionTimeout?
-}
-
 final class GroupDetailsViewControllerSnapshotTests: XCTestCase {
 
     var sut: GroupDetailsViewController!
